pyproject.toml
[build-system]
requires = ["setuptools>=42.0,!=58.*,!=59.*,!=60.*,!=61.*", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdok-geopackage-validator"
description = "Validate Geopackage files according to PDOK requirements and recommendations"
keywords = ["geopackage-validator", "geopackage", "pdok"]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dynamic = ["version", "readme"]
dependencies = [
    "setuptools>=42.0,!=58.*,!=59.*,!=60.*,!=61.*",
    "Click >= 8.0",
    "click-log >=0.3",
    "gdal >=3.4",
    "minio",
    "pydantic >= 2.9",
    "pyyaml",
    "semver >= 3.0.2",
]
requires-python = ">=3.6"

[project.optional-dependencies]
test = [
    "black == 22.3.0",
    "flake8",
    "ipdb",
    "ipython",
    "pytest >= 7.0.1",
    "pytest-cov",
    "pytest-flakes",
    "pytest-mock"
]

[project.scripts]
geopackage-validator = "geopackage_validator.cli:cli"

[tool.setuptools.dynamic]
version = {attr = "geopackage_validator.__version__"}
readme = {file = ["README.md"], content-type = "text/markdown"}

[tool.setuptools.packages.find]
include = ["geopackage_validator*"]  # package names should match these glob patterns (["*"] by default)
exclude = ["tests*"]  # exclude packages matching these glob patterns (empty by default)

[tool.pytest.ini_options]
testpaths = [
    "geopackage_validator",
    "tests",
]